docstore: add a MustDocument method to simplify testing
Avoid having to deal with the error from NewDocument for tests.
